Add ScheduleTest
